+++ src/routes/index.js
@@ -22,6 +22,7 @@ const userActionTypesRouter = require('./user-action-types');
 const offerRouter = require('./offer');
 const notificationChannelsRouter = require('./notification-channels');
 const notificationsRouter = require('./notifications');
+const claimAddressRouter = require('./claim-address');
 
 const router = Router();
 
@@ -47,5 +48,6 @@ router.use('/api/user-action-types', userActionTypesRouter);
 router.use('/api/offer', offerRouter);
 router.use('/api/notification-channels', notificationChannelsRouter);
 router.use('/api/notifications', notificationsRouter);
+router.use('/api/claim-address', claimAddressRouter);
 
 module.exports = router;
